requirements.txt
discord.py>=2.4.0
aiohttp>=3.8.0
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
gunicorn==22.0.0
pillow>=10.0.0
psycopg2-binary>=2.9.0